Added identifier lexing logic
Also added a fix for the gtest library failing to build because warnings
are treated as errors. This was addressed by setting the build
properties for only the `gtest` target to not error on warnings.

I also created a utilities file to hold specific functions related to
string manipulation and scanning such as determining if a character is
an ASCII alphabetic character or not.